Move all actions to lib folder and refactor dashboard route to integrate this move
