+++ app/dashboard/members/[id]/edit/edit-form.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@/app/ui/button';
 import Link from 'next/link';
 import { Member } from '@/app/lib/definitions';
-import { updateMember } from './actions';
+import { updateMember } from '@/app/lib/actions';
 
 // receive the member object as a prop and prefill the form fields with the specific member details
 
@@ -20,8 +20,7 @@ const EditMemberForm = ({ member }: { member: Member }) => {
           name="name"
           readOnly
           defaultValue={member.name}
-          className="input input-bordered input-disabled w-full max-w-xl"
-        />
+          className="input input-bordered input-disabled w-full max-w-xl" />
         {/* Email */}
         <label htmlFor="email">Email</label>
         <input
@@ -29,8 +28,7 @@ const EditMemberForm = ({ member }: { member: Member }) => {
           id="email"
           name="email"
           defaultValue={member.email}
-          className="input input-bordered input-info w-full max-w-xl"
-        />
+          className="input input-bordered input-info w-full max-w-xl" />
         {/* Mobile Number */}
         <label htmlFor="mobileNumber">Mobile Number</label>
         <input
@@ -38,8 +36,7 @@ const EditMemberForm = ({ member }: { member: Member }) => {
           id="mobileNumber"
           name="mobileNumber"
           defaultValue={member.mobileNumber}
-          className="input input-bordered input-info w-full max-w-xl"
-        />
+          className="input input-bordered input-info w-full max-w-xl" />
         {/* County */}
         <label htmlFor="county">County</label>
         <input
@@ -47,8 +44,7 @@ const EditMemberForm = ({ member }: { member: Member }) => {
           id="county"
           name="county"
           defaultValue={member.county}
-          className="input input-bordered input-info w-full max-w-xl"
-        />
+          className="input input-bordered input-info w-full max-w-xl" />
         {/* Constituncy */}
         <label htmlFor="constituency">Constituency</label>
         <input
@@ -56,8 +52,7 @@ const EditMemberForm = ({ member }: { member: Member }) => {
           id="constituency"
           name="constituency"
           defaultValue={member.constituency}
-          className="input input-bordered input-info w-full max-w-xl"
-        />
+          className="input input-bordered input-info w-full max-w-xl" />
         {/* Ward */}
         <label htmlFor="ward">Ward</label>
         <input
@@ -65,14 +60,12 @@ const EditMemberForm = ({ member }: { member: Member }) => {
           id="ward"
           name="ward"
           defaultValue={member.ward}
-          className="input input-bordered input-info w-full max-w-xl"
-        />
+          className="input input-bordered input-info w-full max-w-xl" />
       </div>
       <div className="mt-6 flex justify-end gap-4">
         <Link
           href="/dashboard/members"
-          className="flex h-10 items-center rounded-lg bg-gray-100 px-4 text-sm font-medium text-gray-600 no-underline transition-colors hover:bg-gray-200"
-        >
+          className="flex h-10 items-center rounded-lg bg-gray-100 px-4 text-sm font-medium text-gray-600 no-underline transition-colors hover:bg-gray-200">
           Cancel
         </Link>
         <Button type="submit">Update</Button>
